requirements.txt
tkinter
nltk
textblob
newspaper3k